requirements.txt
# FastAPI Server & Core Framework
fastapi
uvicorn[standard]

# Authentication & Security
python-jose[cryptography]  # JWT token handling
passlib[bcrypt]           # Password hashing
python-multipart          # Form data support
pydantic-settings         # Environment configuration
pydantic[email]           # Email validation

# Core Data & Math Libraries
numpy
pandas
scipy
matplotlib

# Quantitative Analysis Libraries
fbm                # Fractional Brownian Motion
arch               # GARCH models and volatility
hmmlearn           # Hidden Markov Models
mdfa               # Multifractal Detrended Fluctuation Analysis

# Financial Data & APIs
yfinance           # Yahoo Finance API

# AI & LLM Integration
langchain          # LLM framework
langchain-community # Community integrations
scikit-learn       # Machine learning utilities

# Development & Testing
httpx              # HTTP client for testing (FastAPI TestClient)
pytest             # Testing framework (optional)
python-dotenv      # Environment variables

# Optional: Advanced Portfolio Analysis (for Phase 2)
# PyPortfolioOpt   # Portfolio optimization
# backtesting      # Backtesting framework

# Optional: Database (for production)
# sqlalchemy       # Database ORM
# psycopg2-binary  # PostgreSQL adapter
# alembic          # Database migrations